main CI/CD 파이프라인 테스트

+++ src/main/java/com/bbangle/bbangle/review/service/ReviewService.java
@@ -1,42 +1,36 @@
 package com.bbangle.bbangle.review.service;
 
-import static com.bbangle.bbangle.exception.BbangleErrorCode.IMAGE_NOT_FOUND;
-import static com.bbangle.bbangle.exception.BbangleErrorCode.REVIEW_NOT_FOUND;
-
-import com.bbangle.bbangle.review.domain.Badge;
 import com.bbangle.bbangle.board.domain.Board;
 import com.bbangle.bbangle.board.repository.BoardRepository;
 import com.bbangle.bbangle.boardstatistic.service.BoardStatisticService;
 import com.bbangle.bbangle.exception.BbangleErrorCode;
 import com.bbangle.bbangle.exception.BbangleException;
+import com.bbangle.bbangle.image.domain.Image;
+import com.bbangle.bbangle.image.dto.ImageDto;
+import com.bbangle.bbangle.image.repository.ImageRepository;
+import com.bbangle.bbangle.image.service.ImageService;
 import com.bbangle.bbangle.member.domain.Member;
 import com.bbangle.bbangle.member.repository.MemberRepository;
 import com.bbangle.bbangle.page.ImageCustomPage;
 import com.bbangle.bbangle.page.ReviewCustomPage;
-import com.bbangle.bbangle.review.domain.Badge;
-import com.bbangle.bbangle.review.domain.Review;
-import com.bbangle.bbangle.review.domain.ReviewCursor;
-import com.bbangle.bbangle.review.domain.ReviewImg;
-import com.bbangle.bbangle.review.dto.ReviewDto;
-import com.bbangle.bbangle.review.domain.ReviewLike;
-import com.bbangle.bbangle.review.domain.ReviewManager;
-import com.bbangle.bbangle.review.dto.ReviewImagesResponse;
-import com.bbangle.bbangle.review.dto.ReviewImgDto;
-import com.bbangle.bbangle.review.dto.ReviewInfoResponse;
-import com.bbangle.bbangle.review.dto.ReviewRateResponse;
-import com.bbangle.bbangle.review.dto.ReviewRequest;
-import com.bbangle.bbangle.review.dto.SummarizedReviewResponse;
-import com.bbangle.bbangle.review.dto.ReviewSingleDto;
-import com.bbangle.bbangle.review.repository.ReviewImgRepository;
+import com.bbangle.bbangle.review.domain.*;
+import com.bbangle.bbangle.review.dto.*;
 import com.bbangle.bbangle.review.repository.ReviewLikeRepository;
 import com.bbangle.bbangle.review.repository.ReviewRepository;
-import java.math.BigDecimal;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.ObjectUtils;
+
+import java.math.BigDecimal;
+import java.util.*;
+
+import static com.bbangle.bbangle.exception.BbangleErrorCode.IMAGE_NOT_FOUND;
+import static com.bbangle.bbangle.exception.BbangleErrorCode.REVIEW_MEMBER_NOT_PROPER;
+import static com.bbangle.bbangle.exception.BbangleErrorCode.REVIEW_NOT_FOUND;
+import static com.bbangle.bbangle.image.domain.ImageCategory.REVIEW;
+import static java.util.Locale.ROOT;
 
 @Service
 @RequiredArgsConstructor
@@ -44,38 +38,42 @@ public class ReviewService {
 
     private static final Boolean WRITE = true;
     private static final Boolean DELETE = false;
-
     private static final Long PAGE_SIZE = 10L;
     private static final Long NON_MEMBER = 0L;
 
+    @Value("${cdn.domain}")
+    private String cdnDomain;
     private final BoardStatisticService boardStatisticService;
     private final ReviewRepository reviewRepository;
     private final BoardRepository boardRepository;
-    private final ReviewImgRepository reviewImgRepository;
     private final MemberRepository memberRepository;
     private final ReviewLikeRepository reviewLikeRepository;
     private final ReviewManager reviewManager;
     private final ReviewStatistics reviewStatistics;
+    private final ImageService imageService;
+    private final ImageRepository imageRepository;
 
     @Transactional
     public void makeReview(ReviewRequest reviewRequest, Long memberId) {
         memberRepository.findMemberById(memberId);
-        Review review = Review.builder()
-            .content(reviewRequest.content())
-            .rate(reviewRequest.rate())
-            .memberId(memberId)
-            .boardId(reviewRequest.boardId())
-            .build();
+        Review review = Review.of(reviewRequest, memberId);
         List<Badge> badges = reviewRequest.badges();
         badges.forEach(review::insertBadge);
-        reviewRepository.save(review);
-        boardStatisticService.updateReviewCount(reviewRequest.boardId(), review.getRate(), WRITE);
-
-   /*     //FIXME 따로 구현!
-        if(Objects.isNull(reviewRequest.photos())){
+        Review savedReview = reviewRepository.save(review);
+        Long reviewId = savedReview.getId();
+        boardStatisticService.updateReview(reviewRequest.boardId());
+        List<String> urls = reviewRequest.urls();
+        if(Objects.isNull(urls)){
             return;
         }
-        makeReviewImg(reviewRequest, review);*/
+        moveImages(urls, reviewId);
+    }
+
+    @Transactional
+    public ReviewImageUploadResponse uploadReviewImage(ReviewImageUploadRequest reviewImageUploadRequest, Long memberId) {
+        memberRepository.findMemberById(memberId);
+        List<String> urls = imageService.saveAll(reviewImageUploadRequest.category(), reviewImageUploadRequest.images());
+        return new ReviewImageUploadResponse(urls);
     }
 
     @Transactional(readOnly = true)
@@ -92,16 +90,15 @@ public class ReviewService {
         Board board = boardRepository.findById(boardId)
                 .orElseThrow(() -> new BbangleException(BbangleErrorCode.BOARD_NOT_FOUND));
         List<ReviewSingleDto> reviewSingleList = reviewRepository.getReviewSingleList(board.getId(), cursorId);
-        int reviewSingListSize = reviewSingleList.size();
-        Long nextCursor = null;
-        Long lastCursor = null;
-        if(reviewSingListSize != 0){
-            nextCursor = reviewSingleList.get(reviewSingListSize -1).id();
-            lastCursor = reviewSingleList.stream()
-                    .findFirst()
-                    .get()
-                    .id();
+        if(ObjectUtils.isEmpty(reviewSingleList)){
+            return new ReviewCustomPage<>(Collections.emptyList(), 0L, false);
         }
+        int reviewSingListSize = reviewSingleList.size();
+        Long nextCursor = reviewSingleList.get(reviewSingListSize -1).id();
+        Long lastCursor = reviewSingleList.stream()
+                .findFirst()
+                .get()
+                .id();
         ReviewCursor reviewCursor = ReviewCursor.builder()
                 .nextCursor(nextCursor)
                 .lastCursor(lastCursor)
@@ -110,7 +107,7 @@ public class ReviewService {
         if (hasNext) {
             reviewSingleList.remove(reviewSingleList.get(reviewSingListSize - 1));
         }
-        Map<Long, List<ReviewImgDto>> imageMap = reviewRepository.getImageMap(reviewCursor);
+        Map<Long, List<ImageDto>> imageMap = reviewRepository.getImageMap(reviewCursor);
         Map<Long, List<String>> tagMap = new HashMap<>();
         for(ReviewSingleDto reviewSingleDto : reviewSingleList){
             reviewManager.getTagMap(reviewSingleDto, tagMap);
@@ -147,37 +144,28 @@ public class ReviewService {
         ReviewCursor reviewCursor = ReviewCursor.builder()
                 .reviewId(reviewId)
                 .build();
-        Map<Long, List<ReviewImgDto>> imageMap = reviewRepository.getImageMap(reviewCursor);
+        Map<Long, List<ImageDto>> imageMap = reviewRepository.getImageMap(reviewCursor);
         Map<Long, List<String>> tagMap = new HashMap<>();
         reviewManager.getTagMap(reviewDetail, tagMap);
         Map<Long, List<Long>> likeMap = makeLikeMap(reviewCursor);
         return ReviewInfoResponse.create(reviewDetail, imageMap, tagMap, likeMap, memberId);
     }
 
-    private Map<Long, List<Long>> makeLikeMap(ReviewCursor reviewCursor) {
-        List<ReviewLike> likeList = reviewRepository.getLikeList(reviewCursor);
-        return reviewManager.getLikeMap(likeList);
-    }
-
     @Transactional(readOnly = true)
     public ReviewImagesResponse getReviewImages(Long reviewId) {
-        List<ReviewImg> reviewImgs = reviewImgRepository.findByReviewId(reviewId);
-        List<String> urlList = reviewImgs.stream()
-                .map(ReviewImg::getUrl)
-                .toList();
-        return new ReviewImagesResponse(urlList);
+        return new ReviewImagesResponse(imageService.findImagePathById(REVIEW, reviewId));
     }
 
     @Transactional(readOnly = true)
     public ReviewCustomPage<List<ReviewInfoResponse>> getMyReviews(Long memberId, Long cursorId) {
         List<ReviewSingleDto> myReviewList = reviewRepository.getMyReviews(memberId, cursorId);
-        int myReviewListSize = myReviewList.size();
-        Long nextCursor = null;
-        Long lastCursor = null;
-        if(myReviewListSize != 0){
-            nextCursor = myReviewList.get(myReviewListSize -1).id();
-            lastCursor = myReviewList.stream().findFirst().get().id();
+        if(ObjectUtils.isEmpty(myReviewList)){
+            return new ReviewCustomPage<>(Collections.emptyList(), 0L, false);
         }
+        int myReviewListSize = myReviewList.size();
+        Long nextCursor = myReviewList.get(myReviewListSize -1).id();
+        Long lastCursor = myReviewList.stream().findFirst().get().id();
+
         ReviewCursor reviewCursor = ReviewCursor.builder()
                 .nextCursor(nextCursor)
                 .lastCursor(lastCursor)
@@ -186,7 +174,7 @@ public class ReviewService {
         if (hasNext) {
             myReviewList.remove(myReviewList.get(myReviewListSize - 1));
         }
-        Map<Long, List<ReviewImgDto>> imageMap = reviewRepository.getImageMap(reviewCursor);
+        Map<Long, List<ImageDto>> imageMap = reviewRepository.getImageMap(reviewCursor);
         Map<Long, List<String>> tagMap = new HashMap<>();
         for(ReviewSingleDto reviewSingleDto : myReviewList){
             reviewManager.getTagMap(reviewSingleDto, tagMap);
@@ -198,14 +186,14 @@ public class ReviewService {
     }
 
     @Transactional(readOnly = true)
-    public ImageCustomPage<List<ReviewImgDto>> getAllImagesByBoardId(Long boardId, Long cursorId) {
-        List<ReviewImgDto> allImagesByBoardId = reviewRepository.getAllImagesByBoardId(boardId, cursorId);
+    public ImageCustomPage<List<ImageDto>> getAllImagesByBoardId(Long boardId, Long cursorId) {
+        List<ImageDto> allImagesByBoardId = reviewRepository.getAllImagesByBoardId(boardId, cursorId);
+        if(ObjectUtils.isEmpty(allImagesByBoardId)){
+            return new ImageCustomPage<>(Collections.emptyList(), 0L, false);
+        }
         int allImagesSize = allImagesByBoardId.size();
         boolean hasNext = checkHasNext(allImagesSize);
-        Long nextCursor = null;
-        if(allImagesSize != 0){
-            nextCursor = allImagesByBoardId.get(allImagesSize -1).getId();
-        }
+        Long nextCursor = allImagesByBoardId.get(allImagesSize -1).getId();
         if (hasNext) {
             allImagesByBoardId.remove(allImagesByBoardId.get(allImagesSize - 1));
         }
@@ -213,11 +201,11 @@ public class ReviewService {
     }
 
     @Transactional(readOnly = true)
-    public ReviewImgDto getImage(Long imageId) {
-        ReviewImg reviewImg = reviewImgRepository.findById(imageId)
+    public ImageDto getImage(Long imageId) {
+        Image image = imageRepository.findById(imageId)
                 .orElseThrow(() -> new BbangleException(IMAGE_NOT_FOUND));
-        return ReviewImgDto.builder()
-                .url(reviewImg.getUrl())
+        return ImageDto.builder()
+                .url(image.getPath())
                 .build();
     }
     @Transactional
@@ -225,50 +213,76 @@ public class ReviewService {
         memberRepository.findMemberById(memberId);
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new BbangleException(REVIEW_NOT_FOUND));
+        if(!review.getMemberId().equals(memberId)){
+            throw new BbangleException(REVIEW_MEMBER_NOT_PROPER);
+        }
         review.update(reviewRequest);
-        makeReviewImg(reviewRequest, review);
+        List<String> urls = reviewRequest.urls();
+        List<Image> images = imageRepository.findByDomainId(reviewId);
+        List<String> removedUrls = new ArrayList<>();
+        if(!ObjectUtils.isEmpty(images)){
+            List<String> imagePaths = new ArrayList<>(images.stream()
+                    .map(Image::getPath)
+                    .toList());
+            for(String url : urls) {
+                if (imagePaths.contains(url)) {
+                    Image image = imageRepository.findByPath(url);
+                    image.update(reviewId, url);
+                    removedUrls.add(url);
+                    imagePaths.remove(url);
+                }
+            }
+            List<String> changedPaths = removeCdnDomain(imagePaths);
+            imageService.deleteImages(changedPaths);
+            imageRepository.deleteAllByPathIn(imagePaths);
+        }
+        urls = urls.stream()
+                .filter(url -> !removedUrls.contains(url))
+                .toList();
+        if(urls.isEmpty()){
+            return;
+        }
+        moveImages(urls, reviewId);
+    }
+
+    private List<String> removeCdnDomain(List<String> paths) {
+        return paths.stream()
+                .map(path -> path.replace(cdnDomain, ""))
+                .toList();
     }
 
     @Transactional
     public void deleteReview(Long reviewId, Long memberId) {
         memberRepository.findMemberById(memberId);
-        List<ReviewImg> reviewImgs = reviewImgRepository.findByReviewId(reviewId);
-        List<ReviewLike> reviewLikes = reviewLikeRepository.findByReviewId(reviewId);
-        if (!reviewImgs.isEmpty()) {
-            reviewImgRepository.deleteAllInBatch(reviewImgs);
-        }
-        if(!reviewLikes.isEmpty()) {
-            reviewLikeRepository.deleteAllInBatch(reviewLikes);
-        }
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new BbangleException(REVIEW_NOT_FOUND));
+        if(!review.getMemberId().equals(memberId)){
+            throw new BbangleException(REVIEW_MEMBER_NOT_PROPER);
+        }
+        List<Image> reviewImages = imageRepository.findByDomainId(reviewId);
+        List<ReviewLike> reviewLikes = reviewLikeRepository.findByReviewId(reviewId);
+        if (!ObjectUtils.isEmpty(reviewImages)) {
+            imageRepository.deleteAllByDomainId(reviewId);
+        }
+        if(!ObjectUtils.isEmpty(reviewLikes)) {
+            reviewLikeRepository.deleteAllByReviewId(reviewId);
+        }
         review.delete();
-        boardStatisticService.updateReviewCount(review.getBoardId(), review.getRate(), DELETE);
+        boardStatisticService.updateReview(review.getBoardId());
     }
 
     @Transactional
     public void deleteImage(Long imageId) {
-        ReviewImg reviewImg = reviewImgRepository.findById(imageId)
+        Image reviewImg = imageRepository.findById(imageId)
                 .orElseThrow(() -> new BbangleException(IMAGE_NOT_FOUND));
-        reviewImgRepository.delete(reviewImg);
-    }
-
-    private void makeReviewImg(ReviewRequest reviewRequest, Review review) {
-//        Long reviewId = review.getId();
-//        Long boardId = reviewRequest.boardId();
-//        List<String> photos = reviewRequest.photos();
-//        photos.stream()
-//                .map(photo -> s3Service.saveImage(photo, BOARD_FOLDER + boardId + REVIEW_FOLDER + reviewId))
-//                .forEach(reviewImgPath -> reviewImgRepository.save(ReviewImg.builder()
-//                        .reviewId(reviewId)
-//                        .url(reviewImgPath)
-//                        .build()));
+        imageRepository.delete(reviewImg);
     }
 
     private boolean checkHasNext(int size) {
         return size >= PAGE_SIZE + 1;
     }
 
+    @Transactional
     public SummarizedReviewResponse getSummarizedReview(Long boardId) {
         List<ReviewDto> reviews = reviewRepository.findByBoardId(boardId);
 
@@ -282,4 +296,45 @@ public class ReviewService {
 
         return SummarizedReviewResponse.of(averageRatingScore, reviewCount, popularBadgeList);
     }
+
+    private String createNewStoragePath(String extractedFileName, Long reviewId){
+        return REVIEW.name().toLowerCase(ROOT)+"/"+reviewId+extractedFileName;
+    }
+
+    private void moveImages(List<String> urls, Long reviewId){
+        List<Image> images = imageRepository.findAllByPathIn(urls);
+        List<String> fromPaths = makeTempStoragePath(images);
+        List<String> toPaths = makeFinalStoragePath(reviewId, fromPaths);
+        updateImagePath(reviewId, fromPaths, images, toPaths);
+    }
+
+    private void updateImagePath(Long reviewId, List<String> fromPaths, List<Image> images, List<String> toPaths) {
+        List<String> deletedPath = new ArrayList<>();
+        for(int i = 0; i < fromPaths.size(); i++){
+            images.get(i).update(reviewId, cdnDomain + toPaths.get(i));
+            imageService.move(fromPaths.get(i), toPaths.get(i));
+            deletedPath.add(fromPaths.get(i));
+        }
+        imageService.deleteImages(deletedPath);
+    }
+
+    private List<String> makeFinalStoragePath(Long reviewId, List<String> fromPaths) {
+        return fromPaths.stream()
+                .map(path -> path.substring(path.lastIndexOf("/")))
+                .map(extractedFileName -> createNewStoragePath(extractedFileName, reviewId))
+                .toList();
+    }
+
+    private List<String> makeTempStoragePath(List<Image> images) {
+        return images.stream()
+                .map(Image::getPath)
+                .map(path -> path.replace(cdnDomain, ""))
+                .toList();
+    }
+
+    private Map<Long, List<Long>> makeLikeMap(ReviewCursor reviewCursor) {
+        List<ReviewLike> likeList = reviewRepository.getLikeList(reviewCursor);
+        return reviewManager.getLikeMap(likeList);
+    }
+
 }
